add _readline_2.d.ts
